Update the properties file and move things around
