+++ facia-tool/app/controllers/StoriesVisibleController.scala
@@ -67,4 +67,3 @@ object StoriesVisibleController extends Controller {
     }
   }
 }
-
