+++ facia-tool/app/controllers/VanityRedirects.scala
@@ -3,6 +3,7 @@ package controllers
 import model.NoCache
 import play.mvc.Controller
 import play.api.mvc.Results.Redirect
+import auth.PanDomainAuthActions
 
 object VanityRedirects extends Controller with PanDomainAuthActions {
